Split the ARCCHR0003 specs by the branch they exercise
The analyzer has two branches — what a reactor injects, and what it
appends through an injected event store — and every spec for both sat
under when_validating_reactor_dependencies, which describes only the
first. Give the append branch its own when_ folder so the tree reads as
the specification it is.
