+++ Source/DotNET/Chronicle.CodeAnalysis.Specs/for_ReactorEventLogAccessAnalyzer/when_validating_an_append_through_an_event_store/and_reactor_appends_through_a_multi_line_fluent_call.cs
@@ -4,7 +4,7 @@
 using Microsoft.CodeAnalysis;
 using VerifyCS = Cratis.Arc.Chronicle.CodeAnalysis.Specs.Testing.AnalyzerVerifier<Cratis.Arc.Chronicle.CodeAnalysis.ReactorEventLogAccessAnalyzer>;
 
-namespace Cratis.Arc.Chronicle.CodeAnalysis.for_ReactorEventLogAccessAnalyzer.when_validating_reactor_dependencies;
+namespace Cratis.Arc.Chronicle.CodeAnalysis.for_ReactorEventLogAccessAnalyzer.when_validating_an_append_through_an_event_store;
 
 /// <summary>
 /// The reported source text goes into a message the CLI, SARIF readers, and IDE error lists all render on a
