+++ Source/DotNET/Chronicle.CodeAnalysis.Specs/for_ReactorEventLogAccessAnalyzer/when_validating_an_append_through_an_event_store/and_reactor_appends_transactionally_through_event_store_event_log.cs
@@ -4,16 +4,15 @@
 using Microsoft.CodeAnalysis;
 using VerifyCS = Cratis.Arc.Chronicle.CodeAnalysis.Specs.Testing.AnalyzerVerifier<Cratis.Arc.Chronicle.CodeAnalysis.ReactorEventLogAccessAnalyzer>;
 
-namespace Cratis.Arc.Chronicle.CodeAnalysis.for_ReactorEventLogAccessAnalyzer.when_validating_reactor_dependencies;
+namespace Cratis.Arc.Chronicle.CodeAnalysis.for_ReactorEventLogAccessAnalyzer.when_validating_an_append_through_an_event_store;
 
-public class and_reactor_appends_transactionally_to_default_log_through_get_event_sequence : Specification
+public class and_reactor_appends_transactionally_through_event_store_event_log : Specification
 {
     Exception _result;
 
     async Task Because() => _result = await Catch.Exception(async () => await VerifyCS.VerifyAnalyzerAsync(@"
 using System.Threading.Tasks;
 using Cratis.Chronicle;
-using Cratis.Chronicle.EventSequences;
 using Cratis.Chronicle.Reactors;
 
 namespace TestNamespace
@@ -23,13 +22,13 @@ namespace TestNamespace
     public class AuthorNotifier(IEventStore eventStore) : IReactor
     {
         public Task On(AuthorRegistered @event) =>
-            {|#0:eventStore.GetEventSequence(EventSequenceId.Log)|}.Transactional.Append(""some-author"", @event);
+            {|#0:eventStore.EventLog|}.Transactional.Append(""some-author"", @event);
     }
 }",
                 VerifyCS.Diagnostic("ARCCHR0003")
                     .WithSeverity(DiagnosticSeverity.Warning)
                     .WithLocation(0)
-                    .WithArguments("AuthorNotifier", "eventStore.GetEventSequence(EventSequenceId.Log)")));
+                    .WithArguments("AuthorNotifier", "eventStore.EventLog")));
 
     [Fact] void should_report_diagnostic() => _result.ShouldBeNull();
 }
